WriteDiaryEnglishModeStepperをWriteDiaryStepperへ一元化するリファクタリングに関するテスト実装
